Few updates
- Removed Apache Tika dependency as it's too slow for MP3 files - using
mp3agic
- Did implement musicbrainz, however there is a throttling limit and I'm
using parallel streams... waiting a second before I can call it again is
pointless. Will implement it at a later date if I feel there is a
benefit
- Changed the way the UUID is created; for artist there is a chance that
the different artists will be grouped under the same ID because we use
the name. For me, chances of this happening is slim; for albums, we take
the artist and album names and create the UUID that way (even slimmer
that this causes a problem for myself); track uses the artist, album and
track name
- Return UUID of artist and album in the response for tracks. 
- Few more I forget...